Rename command to profile-hermes
